refactor map and chart components
