+++ src/components/chart/dailyCasesChart.js
