+++ src/components/chart/dailyDeathsChart.js
