+++ src/components/chart/rollingAverageCasesChart.js
