+++ src/components/chart/totalCasesChart.js
